move to separate file, reuse vector
