+++ query-planner/Cargo.toml
@@ -9,5 +9,6 @@ edition = "2018"
 
 [dependencies]
 graphql-parser = { path = "../graphql-parser" }
-itertools = "0.9.0"
 indexmap = "1.4.0"
+itertools = "0.9.0"
+lazy_static = "1.4.0"
