+++ query-planner/src/lib.rs
@@ -1,4 +1,8 @@
+#[macro_use]
+extern crate lazy_static;
+
 mod model;
+mod serialize;
 
 #[cfg(test)]
 mod tests {
